[GEN][ZH] Fix compiler warnings for signed/unsigned mismatch of loop indices (#804)

+++ Generals/Code/GameEngine/Source/Common/RTS/SpecialPower.cpp
@@ -225,7 +225,7 @@ SpecialPowerStore::~SpecialPowerStore( void )
 {
 
 	// delete all templates
-	for( Int i = 0; i < m_specialPowerTemplates.size(); ++i )
+	for( size_t i = 0; i < m_specialPowerTemplates.size(); ++i )
 		deleteInstance(m_specialPowerTemplates[ i ]);
 
 	// erase the list
@@ -242,7 +242,7 @@ SpecialPowerTemplate* SpecialPowerStore::findSpecialPowerTemplatePrivate( AsciiS
 {
 
 	// search the template list for matching name
-	for( Int i = 0; i < m_specialPowerTemplates.size(); ++i )
+	for( size_t i = 0; i < m_specialPowerTemplates.size(); ++i )
 		if( m_specialPowerTemplates[ i ]->getName() == name )
 			return m_specialPowerTemplates[ i ];
 
@@ -257,7 +257,7 @@ const SpecialPowerTemplate *SpecialPowerStore::findSpecialPowerTemplateByID( Uns
 {
 
 	// search the template list for matching name
-	for( Int i = 0; i < m_specialPowerTemplates.size(); ++i )
+	for( size_t i = 0; i < m_specialPowerTemplates.size(); ++i )
 		if( m_specialPowerTemplates[ i ]->getID() == id )
 			return m_specialPowerTemplates[ i ];
