+++ Generals/Code/GameEngine/Source/GameClient/GUI/ControlBar/ControlBarCommand.cpp
@@ -364,7 +364,7 @@ void ControlBar::populateCommand( Object *obj )
 							//button specifying a vector of sciences in the command button.
 							Int bestIndex = -1;
 							ScienceType science;
-							for( Int scienceIndex = 0; scienceIndex < commandButton->getScienceVec().size(); ++scienceIndex )
+							for( size_t scienceIndex = 0; scienceIndex < commandButton->getScienceVec().size(); ++scienceIndex )
 							{
 								science = commandButton->getScienceVec()[ scienceIndex ];
