+++ Generals/Code/GameEngine/Source/GameLogic/Object/Locomotor.cpp
@@ -2706,7 +2706,7 @@ void LocomotorSet::xferSelfAndCurLocoPtr(Xfer *xfer, Locomotor** loco)
 		}
 		else
 		{
-			for (int i = 0; i < m_locomotors.size(); ++i)
+			for (size_t i = 0; i < m_locomotors.size(); ++i)
 			{
 				if (m_locomotors[i]->getTemplateName() == name)
 				{
@@ -2724,7 +2724,7 @@ void LocomotorSet::xferSelfAndCurLocoPtr(Xfer *xfer, Locomotor** loco)
 //-------------------------------------------------------------------------------------------------
 void LocomotorSet::clear()
 {
-	for (int i = 0; i < m_locomotors.size(); ++i)
+	for (size_t i = 0; i < m_locomotors.size(); ++i)
 	{
 		if (m_locomotors[i])
 			deleteInstance(m_locomotors[i]);
