+++ Generals/Code/GameEngine/Source/GameLogic/System/CrateSystem.cpp
@@ -152,7 +152,7 @@ CrateTemplate *CrateSystem::newCrateTemplateOverride( CrateTemplate *crateToOver
 const CrateTemplate *CrateSystem::findCrateTemplate(AsciiString name) const
 {
 	// search weapon list for name
-	for (Int i = 0; i < m_crateTemplateVector.size(); i++)
+	for (size_t i = 0; i < m_crateTemplateVector.size(); i++)
 		if(m_crateTemplateVector[i]->getName() == name) {
 			CrateTemplateOverride overridable(m_crateTemplateVector[i]);
 			return overridable;
@@ -165,7 +165,7 @@ const CrateTemplate *CrateSystem::findCrateTemplate(AsciiString name) const
 CrateTemplate *CrateSystem::friend_findCrateTemplate(AsciiString name)
 {
 	// search weapon list for name
-	for (Int i = 0; i < m_crateTemplateVector.size(); i++)
+	for (size_t i = 0; i < m_crateTemplateVector.size(); i++)
 		if(m_crateTemplateVector[i]->getName() == name) {
 			CrateTemplateOverride overridable(m_crateTemplateVector[i]);
 			return const_cast<CrateTemplate*>((const CrateTemplate *)overridable);
